[GH-165] Better chain_state validation
Instead of calculating chain state of a block and then combining it with
old chain state we take old chain_state and apply transactions to it.
While applying transactions we do all requrired checks. One of them is
missing previously nonce check.

Overall the code is simplified and redundand calculations were removed